Update Documentation To Not Use persist Method
Update documentation for the persisting of the repository to use the
create method instead of the persist method due to the persist method
not existing for quite some time.

The example could confuse a user trying to implement the pattern causing
false issues being posted or time loss.

Relates to issue #213